+++ p_mobj.h
@@ -321,7 +321,7 @@ boolean P_SetMobjState(mobj_t __far* mobj, statenum_t state);
 void    P_MobjThinker(mobj_t __far* mobj);
 
 void    P_SpawnPuff(fixed_t x, fixed_t y, fixed_t z);
-void    P_SpawnBlood(fixed_t x, fixed_t y, fixed_t z, int32_t damage);
+void    P_SpawnBlood(fixed_t x, fixed_t y, fixed_t z, int16_t damage);
 mobj_t __far* P_SpawnMissile(mobj_t __far* source, mobj_t __far* dest, mobjtype_t type);
 void    P_SpawnPlayerMissile(mobj_t __far* source, mobjtype_t type);
 void    P_SpawnMapThing (const mapthing_t __far* mthing);
